update build system
