+++ agent/pom.xml
@@ -1,0 +1,101 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>org.iser21</groupId>
+    <artifactId>jmtrace-agent</artifactId>
+    <version>1.0</version>
+
+    <properties>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
+        <asm.version>8.0.1</asm.version>
+    </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <artifactId>asm</artifactId>
+                <groupId>org.ow2.asm</groupId>
+                <version>${asm.version}</version>
+            </dependency>
+
+            <dependency>
+                <artifactId>asm-tree</artifactId>
+                <groupId>org.ow2.asm</groupId>
+                <version>${asm.version}</version>
+            </dependency>
+
+            <dependency>
+                <artifactId>asm-analysis</artifactId>
+                <groupId>org.ow2.asm</groupId>
+                <version>${asm.version}</version>
+            </dependency>
+
+            <dependency>
+                <artifactId>asm-commons</artifactId>
+                <groupId>org.ow2.asm</groupId>
+                <version>${asm.version}</version>
+            </dependency>
+
+            <dependency>
+                <artifactId>asm-util</artifactId>
+                <groupId>org.ow2.asm</groupId>
+                <version>${asm.version}</version>
+            </dependency>
+
+            <dependency>
+                <artifactId>asm-xml</artifactId>
+                <groupId>org.ow2.asm</groupId>
+                <version>${asm.version}</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+    
+    <dependencies>
+        <dependency>
+            <artifactId>asm-commons</artifactId>
+            <groupId>org.ow2.asm</groupId>
+            <version>${asm.version}</version>
+        </dependency>
+
+        <dependency>
+            <artifactId>asm-util</artifactId>
+            <groupId>org.ow2.asm</groupId>
+            <version>${asm.version}</version>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-shade-plugin</artifactId>
+                <version>3.0.0</version>
+                <executions>
+                    <execution>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>shade</goal>
+                        </goals>
+                        <configuration>
+                            <transformers>
+                                <transformer
+                                        implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
+                                    <manifestEntries>
+                                        <premain-class>iser21.jmtrace.JMTraceAgent</premain-class>
+                                        <Can-Redefine-Classes>true</Can-Redefine-Classes>
+                                        <Can-Retransform-Classes>true</Can-Retransform-Classes>
+                                    </manifestEntries>
+                                </transformer>
+                            </transformers>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
+
+</project>
