+++ example/src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Created-By: 11.0.12 (Oracle Corporation)
+Main-Class: example.HeavySortMain
