--- src/main/resources/META-INF/MANIFEST.MF
@@ -1,4 +1,0 @@
-Manifest-Version: 1.0
-Premain-Class: iser21.jmtrace.JMTraceAgent
-Can-Redefine-Classes: true
-Can-Retransform-Classes: true
