--- src/test/resources/example/MANIFEST.MF
@@ -1,4 +1,0 @@
-Manifest-Version: 1.0
-Created-By: 11.0.12 (Oracle Corporation)
-Main-Class: example.HeavySortMain
-
